// src/core/types/DocumentSecurityFeature.ts
export enum DocumentSecurityFeature {
  IR_A = "IR_A",
  IR_B = "IR_B",
  IR_C = "IR_C",
  IR_D = "IR_D",
  IR_E = "IR_E",
  IR_F = "IR_F",
  OFFSET_DESIGN_A = "OFFSET_DESIGN_A",
  OFFSET_DESIGN_B = "OFFSET_DESIGN_B",
  OFFSET_DESIGN_C = "OFFSET_DESIGN_C",
  OFFSET_DESIGN_D = "OFFSET_DESIGN_D",
  OFFSET_DESIGN_E = "OFFSET_DESIGN_E",
  OFFSET_DESIGN_F = "OFFSET_DESIGN_F",
  OFFSET_DESIGN_G = "OFFSET_DESIGN_G",
  OFFSET_DESIGN_H = "OFFSET_DESIGN_H",
}